+++ Backend/config/db.ts
@@ -1,14 +1,14 @@
-import mongoose from "mongoose"
+import mongoose from "mongoose";
 
 const connectDB = () => {
-    mongoose.connect(process.env.DB_URL||"")
+  mongoose
+    .connect(process.env.DB_URL || "")
     .then(() => {
-        console.log("MongoDB is Connected...");
+      console.log("MongoDB is Connected...");
     })
-    .catch(err => {
-        console.error('DB connection failed:', err);
+    .catch((err) => {
+      console.error("DB connection failed:", err);
+    });
+};
 
-    })
-}
-
-export default connectDB
+export default connectDB;
