+++ Backend/routes/mcq.ts
@@ -1,0 +1,8 @@
+import express from "express";
+import mcqController from "../controller/mcq";
+
+const router = express.Router();
+
+router.get("/", mcqController);
+
+export default router;
