+++ Backend/routes/summary.ts
@@ -1,0 +1,8 @@
+import express from "express";
+import summaryController from "../controller/summary";
+
+const router = express.Router();
+
+router.get("/", summaryController);
+
+export default router;
